Moving some functions out to a common python file and adjusted the instruction pattern and pair detection scripts to fit with this
